Move menu updated to utils
