+++ projects/packages/forms/src/dashboard/inbox/dataviews/actions.js
@@ -7,24 +7,7 @@ import { store as noticesStore } from '@wordpress/notices';
 import { notSpam, spam } from '../../icons';
 import { store as dashboardStore } from '../../store';
 import InboxResponse from '../response';
-/**
- * Update the unread count in the admin menu.
- *
- * @param {number} count - The new unread count.
- */
-export const updateMenuCounter = count => {
-	// iterate over all elements with the class 'jp-feedback-unread-counter' and update their text content
-	document.querySelectorAll( '.jp-feedback-unread-counter' ).forEach( item => {
-		if ( item.dataset.unreadDiff ) {
-			const newCount = parseInt( item.dataset.unreadDiff, 10 ) + count;
-			item.textContent = newCount > 0 ? newCount : '';
-			item.style.display = newCount > 0 ? '' : 'none';
-		} else {
-			item.textContent = count > 0 ? count : '';
-			item.style.display = count > 0 ? '' : 'none';
-		}
-	} );
-};
+import { updateMenuCounter } from '../utils';
 
 export const BULK_ACTIONS = {
 	markAsSpam: 'mark_as_spam',
